+++ plugin/src/main/groovy/org/openbakery/simulators/SimulatorsStartTask.groovy
@@ -2,6 +2,7 @@ package org.openbakery.simulators
 
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.TaskAction
+import org.openbakery.CommandRunner
 import org.openbakery.Type
 import org.openbakery.XcodePlugin
 
@@ -11,7 +12,7 @@ class SimulatorsStartTask extends DefaultTask {
 	public SimulatorsStartTask() {
 		setDescription("Start iOS Simulators")
 		dependsOn(XcodePlugin.XCODE_BUILD_TASK_NAME)
-		simulatorControl = new SimulatorControl(project)
+		simulatorControl = new SimulatorControl(project, new CommandRunner())
 	}
 
 	@TaskAction