+++ plugin/src/test/groovy/org/openbakery/simulators/SimulatorsRunAppTaskSpecification.groovy
@@ -8,6 +8,7 @@ import org.openbakery.XcodePlugin
 import org.openbakery.stubs.PlistHelperStub
 import org.junit.Before
 import org.junit.Test
+import spock.lang.Specification
 
 import static org.hamcrest.MatcherAssert.assertThat
 import static org.hamcrest.Matchers.*
@@ -15,15 +16,14 @@ import static org.hamcrest.Matchers.*
 /**
  * Created by rene on 01.09.15.
  */
-class SimulatorsRunAppTaskTest {
+class SimulatorsRunAppTaskSpecification extends Specification {
 
 	SimulatorsRunAppTask task
 	Project project
 	File projectDir
 
 
-	@Before
-	void setup() {
+	def "setup"() {
 
 		projectDir = new File(System.getProperty("java.io.tmpdir"), "gradle-xcodebuild")
 		project = ProjectBuilder.builder().withProjectDir(projectDir).build()
@@ -34,52 +34,62 @@ class SimulatorsRunAppTaskTest {
 	}
 
 
-	@Test
-	void create() {
-		assert task instanceof SimulatorsRunAppTask
-		assert task.simulatorControl instanceof SimulatorControl
+	def create() {
+		expect:
+		task instanceof SimulatorsRunAppTask
+		task.simulatorControl instanceof SimulatorControl
 	}
 
 
-	@Test(expected = IllegalArgumentException)
-	void hasNoInfoPlist() {
+	def "hasNoInfoPlist"() {
+		when:
 		task.run()
+
+		then:
+		thrown(IllegalArgumentException.class)
+
 	}
 
 
-	@Test(expected = IllegalArgumentException)
-	void noSimulatorSDK() {
+	def "no simulator SDK"() {
+		given:
 		project.xcodebuild.infoPlist =  "Info.plist"
 		project.xcodebuild.type = Type.OSX
+		when:
 		task.run()
-	}
 
-	@Test(expected = IllegalArgumentException)
-	void bundleIdentifierIsNull() {
-		project.xcodebuild.infoPlist =  "Info.plist"
-
-		task.run()
+		then:
+		thrown(IllegalArgumentException.class)
 
 	}
 
-	@Test
-	void run() {
+	def "bundleIdentifier is null"() {
+		given:
 		project.xcodebuild.infoPlist =  "Info.plist"
 
-		def mock = new MockFor(SimulatorControl)
-		mock.demand.simctl { p1, p2, p3 ->
-			assertThat(p1, is("launch"));
-			assertThat(p2, is("booted"));
-			assertThat(p3, is("com.example.Example"));
-		}
+		when:
+		task.run()
 
-		task.simulatorControl = mock.proxyInstance()
+		then:
+		thrown(IllegalArgumentException.class)
+	}
+
+	def "run"() {
+		given:
+		project.xcodebuild.infoPlist =  "Info.plist"
+
+		SimulatorControl simulatorControl = Mock(SimulatorControl)
+		task.simulatorControl = simulatorControl
+
 		task.plistHelper = new PlistHelperStub([
 						CFBundleIdentifier: "com.example.Example"
 		])
+
+		when:
 		task.run()
 
-		mock.verify task.simulatorControl
+		then:
+		1 * simulatorControl.simctl("launch", "booted", "com.example.Example")
 	}